refactor: depend on mrpunyapal/git-reader ^0.1 — extract read-only Git client to standalone package
