+++ composer.json
@@ -6,6 +6,7 @@
     "require": {
         "php": "^8.3",
         "league/commonmark": "^2.7",
+        "mrpunyapal/git-reader": "^0.1",
         "phiki/phiki": "^2.2"
     },
     "require-dev": {
