--- src/Hub/Git/GitException.php
@@ -1,11 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Docsmith\Hub\Git;
-
-use RuntimeException;
-
-class GitException extends RuntimeException
-{
-}
