--- src/Hub/Git/RepositoryNotFoundException.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Docsmith\Hub\Git;
-
-final class RepositoryNotFoundException extends GitException
-{
-}
